backend/requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.6
pydantic==2.8.2
python-multipart==0.0.9
pypdf==4.2.0
langchain==0.2.12
langchain-community==0.2.11
langchain-openai==0.1.21
chromadb==0.5.5
sentence-transformers==3.0.1
python-dotenv==1.0.1
tiktoken==0.7.0
